[IMP] marin: add tokenization logic to infer account in invoice lines
Add logic to _compute_account_id that attempts to assign an account based on
tokenized keywords from the line's name. This applies to invoice lines without
a product but with a partner. It matches against non-deprecated accounts of
type 'expense' or 'income' based on move type.

Fallback to most frequent partner account remains if no suitable match is found.

Also added comments to code for clarity and consistency, according to guidelines.